--- Minishell/srcs/ms_free_args.c
@@ -1,20 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ms_pipe_list_utils.c                               :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: maliew <maliew@student.42kl.edu.my>        +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/10/01 16:10:09 by maliew            #+#    #+#             */
-/*   Updated: 2022/10/10 11:10:39 by maliew           ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "minishell.h"
-
-void	ms_free_args(void *content)
-{
-	if (*(char **)content)
-		free(*(char **)content);
-	free(content);
-}
